финальные добавления хедера
- исправляет обрезание сайдбара на телефоне;
- добавляет якорь на кнопу перехода отдельно к форме и к контакатам;
- помещает сайдбар в бекдроп, меняет стили бекдропа для таблетки и десктопа;
- добавляет ариа-лейблы к ссылкам;
- добавляет иконку с якорем возврата к хедеру;
- добавляет анимацию перехода к якорю;
