+++ src/index.html
@@ -14,6 +14,7 @@
     />
   </head>
   <body>
+    <include src="./partials/header.html"></include>
     <include src="./partials/slider.html"></include>
     <include src="./partials/about.html"></include>
     <include src="./partials/price.html"></include>
@@ -25,7 +26,5 @@
 
     <!-- Скрипт идёт последним в разметке. В нём уже подключен главный SASS-файл. -->
     <script src="./index.js"></script>
-    <script src="./js/menu.js"></script>
-    <script src="./js/header.js"></script>
   </body>
 </html>
