+++ src/partials/contacts.html
@@ -1,5 +1,5 @@
 <section class="section section--background">
-  <form class="recording container">
+  <form class="recording container" id="form">
     <h2 class="section-title recording-header">Онлайн запись</h2>
     <ul class="recording-list">
       <li class="recording-item">
@@ -36,7 +36,7 @@
     </ul>
     <button type="submit" class="button colorReverse">Отправить</button>
   </form>
-  <div class="contacts">
+  <div class="contacts" id="contacts">
     <div class="container desktop-container">
       <h2 class="section-title recording-header">Контакты</h2>
       <div class="contacts-flex">
