+++ src/partials/header.html
@@ -1,4 +1,4 @@
-<header class="header">
+<header class="header" id="top">
   <div class="headerWrap">
     <div class="headerMobile">
       <svg class="logo" width="69" height="56">
@@ -7,45 +7,88 @@
           aria-label="Логотип компании"
         ></use>
       </svg>
-      <button class="buttonMenu" aria-label="кнопка меню" data-menu-button>
+      <button
+        class="buttonMenu"
+        aria-label="кнопка меню открыть-закрыть"
+        aria-expanded="false"
+        aria-controls="sideBar"
+        data-menu-button
+      >
         <svg width="40" height="40">
           <use href="./images/sprite.svg#icon-menu" class="icon-menu"></use>
           <use href="./images/sprite.svg#icon-cross" class="icon-cross"></use>
         </svg>
       </button>
     </div>
-    <div class="backdrop" data-menu></div>
-    <div class="sideBar" data-sideBar>
-      <nav>
-        <ul class="siteNav">
-          <li class="siteNav-item">
-            <a href="#aboutUs" class="siteNav-link">О нас</a>
-          </li>
-          <li class="siteNav-item">
-            <a href="#price" class="siteNav-link">Услуги и цены</a>
-          </li>
-          <li class="siteNav-item">
-            <a href="#team" class="siteNav-link">Мастера</a>
-          </li>
-          <li class="siteNav-item">
-            <a href="#contacts" class="siteNav-link">Контакты</a>
-          </li>
-        </ul>
-      </nav>
-      <div class="headerContacts">
-        <a href="tel:+380441111111" class="tel">+38 044 111 11 11</a>
-        <button class="button colorReverse-header">онлайн-запись</button>
-      </div>
-      <div class="socialWrap">
-        <ul class="headerSocial">
-          <li class="headerSocial-item">
-            <a href="" class="headerSocial-link">Instagram</a>
-          </li>
-          <li class="headerSocial-item">
-            <a href="" class="headerSocial-link">Youtube</a>
-          </li>
-        </ul>
+    <div class="backdrop" data-menu>
+      <div class="sideBar" id="sideBar" data-sideBar>
+        <nav>
+          <ul class="siteNav">
+            <li class="siteNav-item">
+              <a
+                href="#aboutUs"
+                class="siteNav-link"
+                aria-label="кнопка перехода к секции о нас"
+                data-menu-close-1
+                >О нас</a
+              >
+            </li>
+            <li class="siteNav-item">
+              <a
+                href="#price"
+                class="siteNav-link"
+                aria-label="кнопка перехода к секции цены"
+                data-menu-close-2
+                >Услуги и цены</a
+              >
+            </li>
+            <li class="siteNav-item">
+              <a
+                href="#team"
+                class="siteNav-link"
+                aria-label="кнопка перехода к секции наша команда"
+                data-menu-close-3
+                >Мастера</a
+              >
+            </li>
+            <li class="siteNav-item">
+              <a
+                href="#contacts"
+                class="siteNav-link"
+                aria-label="кнопка перехода к контактам"
+                data-menu-close-4
+                >Контакты</a
+              >
+            </li>
+          </ul>
+        </nav>
+        <div class="headerContacts">
+          <a href="tel:+380441111111" class="tel">+38 044 111 11 11</a>
+          <a
+            href="#form"
+            class="button colorReverse-header"
+            aria-label="кнопка перехода на форму записи"
+            data-menu-close-5
+          >
+            онлайн-запись
+          </a>
+        </div>
+        <div class="socialWrap">
+          <ul class="headerSocial">
+            <li class="headerSocial-item">
+              <a href="" class="headerSocial-link">Instagram</a>
+            </li>
+            <li class="headerSocial-item">
+              <a href="" class="headerSocial-link">Youtube</a>
+            </li>
+          </ul>
+        </div>
       </div>
     </div>
+    <a href="#top" class="iconBack" aria-label="кнопка вернуться на верх">
+      <svg width="32" height="32">
+        <use href="./images/sprite.svg#icon-circle-up"></use>
+      </svg>
+    </a>
   </div>
 </header>
