+++ src/partials/slider.html
@@ -1,5 +1,4 @@
 <section class="slider">
-  <include src="./partials/header.html"></include>
   <div class="siema">
     <picture>
       <source
