+++ src/sass/components/_slider.scss
@@ -1,6 +1,5 @@
 .slider {
   position: relative;
-  overflow: hidden;
 
   max-width: 1600px;
   margin-left: auto;
@@ -144,5 +143,4 @@
   @media screen and (min-width: $desktop) {
     left: -35px;
   }
-
 }
